efikamx: port the USB support to iomux-v3
* Move from the old iomux API to the new one.
* Correct the comment about the USB "violation" and attempt to explain
  why this even exists. If anyone has any questions on this I can walk
  them through the schematics if necessary.
* efika -> efikamx, any strings "Efika" to "Efika MX" for consistency.
* gpio_direction_output was being called a few times on a few GPIOs, so
  only ever set the direction once (in efikamx_setup_usb), and just
  gpio_set_value after that, also for consistency.

Built and booted on an Efika Smarttop and Smartbook for verification.

USB seems not to work correctly on Smartbook, but without this patch it has
the same behavior, and as far as we know it never worked on the Efika Smartbook
- so we have some bugs here that this patch makes NO attempt to fix.

I have verified all the IOMUX settings and code path is the same before and
after the patch, with the same effect on the hardware and logic levels on
the board as the old code. On Smarttop, only Ethernet (DR port) was tested,
and it worked fine. Further testing will be required on the USBH1 port.

A version of this patch got lost in a previous upstream attempt, so we're
submitting it to lay some groundwork to fixing the actual bugs here since it
is a) quite odd for efikamx.c to use iomux-mx51.h and efikamx-usb.c to use
iomux.h and b) it will make for a less clumsy patch later.

Signed-off-by: Matt Sealey <matt@genesi-usa.com>