efikamx: update configuration
This patch supplies a clean, readable and updated configuration for
Efika MX Smarttop and Smartbook systems.

Mainly this implements the "scan for boot.scr" bootcmd similar to the
one used in "Production" Efika MX U-Boot.

Other than that, it just cleans everything up, reorders things for
readability and adds lots of comments about the reasons for the way
some of the configuration is laid out. This configuration has been
tested in the lab on many, many example devices for about 6 months now,
keeping track of upstream, and we feel it is ready to go live as it
stands.

For developers wanting to use this instead of "Production" shipped
U-Boot, the following major differences will be noticed (this also
applies to people who used the previous upstream config);

* Boot Delay is now 1 second instead of 3, to reduce boot times.

* It will no longer support a gzip-compressed uImage - but this scenario
  is basically not used in real life anyway (all reasonable Efika MX
  kernels and support files were created with 'mkimage -C none ...')

* You cannot save your U-Boot environment with this config. Please
  use boot.scr to change anything that needs changing or in extreme
  cases you will need to rebuild U-Boot (but anyone reading this log
  is perfectly capable of that).

There are lots of other little tweaks which basically align it with
Production and disable features nobody ever used.

Signed-off-by: Matt Sealey <matt@genesi-usa.com>