ubifs: enable disabling of compression types
Without CONFIG_LZO or CONFIG_GZIP (and dependent CONFIG_ZLIB) ubifs
will not build - however it is possible we do not want to support
these compression types on a live NAND filesystem to save space in
the U-Boot binary.

Wrap all uses of zlib/lzo symbols in ubifs with relevant preprocessor
macros to enable it to be compiled out by #undef CONFIG_LZO etc.

Signed-off-by: Matt Sealey <matt@genesi-usa.com>