efikamx: add NAND support for Slimbook
Not working yet, NAND debug is on and will print out what is
going on, but it does not seem to detect any actual NAND on
the bus. This may be.. voltage, clock, some kind of internal
NAND controller bits need setting to unlock it, maybe, since
Linux does something similar. Will need to check if this is
being done in U-Boot since MX53 NAND support is pretty fresh.

Signed-off-by: Matt Sealey <matt@genesi-usa.com>